2004-12-09 12:25 UTC+0300 Alexander Kresin <alex@belacy.belgorod.su>
git-svn-id: https://svn.code.sf.net/p/hwgui/code/trunk/hwgui@379 66345640-f411-406e-acba-706ed0d45c9c

+++ source/procmisc/arr2str.c
@@ -1,5 +1,5 @@
 /*
- * $Id: arr2str.c,v 1.4 2004-12-09 13:45:37 alkresin Exp $
+ * $Id: arr2str.c,v 1.5 2004-12-10 09:25:04 alkresin Exp $
  *
  * HWGUI - Harbour Win32 GUI library source code:
  * Array / String conversion functions
@@ -40,7 +40,8 @@ static char * ReadArray( char * ptr, PHB_ITEM pItem )
          unsigned int iLen;
          ptr ++;
          // iLen = ( *ptr + ( ( *(ptr+1) ) << 8 ) ) & 0xffff;
-         iLen = *ptr & 0x00ff + ( ( *(ptr+1) ) << 8 ) & 0xffff;
+         iLen = ( (unsigned int)*ptr & 0x00ff ) + 
+            ( (unsigned int)( ( *(ptr+1) ) << 8 ) ) & 0xffff;
          ptr ++; ptr ++;
          temp = hb_itemPutCL( NULL, ptr, iLen );
          ptr += iLen;
@@ -81,8 +82,10 @@ static char * ReadArray( char * ptr, PHB_ITEM pItem )
       {
          unsigned int iLen;
          ptr ++;
-         iLen = ( *ptr & 0x00ff + ( ( *(ptr+1) ) << 8 ) & 0xff00 + 
-               ( ( *(ptr+2) ) << 16 ) & 0xff0000 + ( ( *(ptr+3) ) << 24 ) ) & 0xff000000;
+         iLen = ( (unsigned int)*ptr & 0x00ff ) + 
+                ( (unsigned int)( ( *(ptr+1) ) << 8 ) ) & 0xff00 + 
+                ( (unsigned int)( ( *(ptr+2) ) << 16 ) ) & 0xff0000 + 
+                ( (unsigned int)( ( *(ptr+3) ) << 24 ) ) & 0xff000000;
          ptr ++; ptr ++; ptr ++; ptr ++;
          temp = hb_itemPutCL( NULL, ptr, iLen );
          ptr += iLen;
